@@ -1,16 +1,17 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.FileProviders.Embedded" Version="6.0.36" />
-    <PackageReference Include="QuestPDF" Version="2021.9.3" />
-    <PackageReference Include="SkiaSharp.NativeAssets.Linux" Version="2.88.3" />
+    <PackageReference Include="Microsoft.Extensions.FileProviders.Embedded" Version="9.0.6" />
+    <PackageReference Include="QuestPDF" Version="2025.5.1" />
+    <PackageReference Include="SixLabors.ImageSharp" Version="3.1.10" />
+    <PackageReference Include="SkiaSharp.NativeAssets.Linux" Version="3.119.0" />
     <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
-    <PackageReference Include="Barcoder" Version="1.0.0.90" />
-    <PackageReference Include="Barcoder.Renderer.Image" Version="1.0.0.90" />
+    <PackageReference Include="Barcoder" Version="3.0.0" />
+    <PackageReference Include="Barcoder.Renderer.Image" Version="3.0.0" />
   </ItemGroup>
 
   <ItemGroup>
